Revert "refactor: ProgrammeMembership entity"

+++ src/main/java/uk/nhs/hee/tis/trainee/sync/service/ProgrammeMembershipSyncService.java
@@ -24,15 +24,11 @@ package uk.nhs.hee.tis.trainee.sync.service;
 import static uk.nhs.hee.tis.trainee.sync.model.Operation.DELETE;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import java.time.LocalDate;
-import java.util.Collections;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
-import java.util.UUID;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
-import uk.nhs.hee.tis.trainee.sync.mapper.ProgrammeMembershipMapper;
 import uk.nhs.hee.tis.trainee.sync.model.ProgrammeMembership;
 import uk.nhs.hee.tis.trainee.sync.model.Record;
 import uk.nhs.hee.tis.trainee.sync.repository.ProgrammeMembershipRepository;
@@ -46,75 +42,73 @@ public class ProgrammeMembershipSyncService implements SyncService {
   private final DataRequestService dataRequestService;
 
   private final RequestCacheService requestCacheService;
-  private final ProgrammeMembershipMapper mapper;
 
   ProgrammeMembershipSyncService(ProgrammeMembershipRepository repository,
-      DataRequestService dataRequestService, RequestCacheService requestCacheService,
-      ProgrammeMembershipMapper mapper) {
+      DataRequestService dataRequestService, RequestCacheService requestCacheService) {
     this.repository = repository;
     this.dataRequestService = dataRequestService;
     this.requestCacheService = requestCacheService;
-    this.mapper = mapper;
   }
 
   @Override
-  public void syncRecord(Record programmeMembershipRecord) {
-    ProgrammeMembership programmeMembership = mapper.toEntity(programmeMembershipRecord.getData());
+  public void syncRecord(Record programmeMembership) {
+    if (!(programmeMembership instanceof ProgrammeMembership)) {
+      String message = String.format("Invalid record type '%s'.", programmeMembership.getClass());
+      throw new IllegalArgumentException(message);
+    }
 
-    if (programmeMembershipRecord.getOperation().equals(DELETE)) {
-      repository.deleteById(programmeMembership.getUuid());
+    if (programmeMembership.getOperation().equals(DELETE)) {
+      repository.deleteById(programmeMembership.getTisId());
     } else {
-      repository.save(programmeMembership);
+      repository.save((ProgrammeMembership) programmeMembership);
     }
 
     requestCacheService.deleteItemFromCache(ProgrammeMembership.ENTITY_NAME,
-        programmeMembership.getUuid().toString());
+        programmeMembership.getTisId());
   }
 
-  public Optional<ProgrammeMembership> findById(String uuid) {
-    return repository.findById(UUID.fromString(uuid));
+  public Optional<ProgrammeMembership> findById(String id) {
+    return repository.findById(id);
   }
 
   public Set<ProgrammeMembership> findByProgrammeId(String programmeId) {
-    return repository.findByProgrammeId(Long.parseLong(programmeId));
+    return repository.findByProgrammeId(programmeId);
   }
 
   public Set<ProgrammeMembership> findByCurriculumId(String curriculumId) {
-    // TODO: get PMs for curriculum ID.
-    return Collections.emptySet();
+    return repository.findByCurriculumId(curriculumId);
   }
 
   public Set<ProgrammeMembership> findByPersonId(String personId) {
-    return repository.findByPersonId(Long.parseLong(personId));
+    return repository.findByPersonId(personId);
   }
 
-  public Set<ProgrammeMembership> findBySimilar(String personId, String programmeId,
-      String programmeMembershipType, String programmeStartDate, String programmeEndDate) {
-    return repository.findBySimilar(Long.parseLong(personId), Long.parseLong(programmeId),
-        programmeMembershipType,
-        LocalDate.parse(programmeStartDate), LocalDate.parse(programmeEndDate));
+  public Set<ProgrammeMembership> findBySimilar(String personId,
+      String programmeId,
+      String programmeMembershipType,
+      String programmeStartDate,
+      String programmeEndDate) {
+    return repository.findBySimilar(personId, programmeId, programmeMembershipType,
+        programmeStartDate, programmeEndDate);
   }
 
   /**
-   * Make a request to retrieve a specific programme membership.
+   * Make a request to retrieve a specific post.
    *
-   * @param uuid The uuid of the post to be retrieved.
+   * @param id The id of the post to be retrieved.
    */
-  public void request(UUID uuid) {
-    String uuidString = uuid.toString();
-
-    if (!requestCacheService.isItemInCache(ProgrammeMembership.ENTITY_NAME, uuidString)) {
-      log.info("Sending request for ProgrammeMembership [{}]", uuidString);
+  public void request(String id) {
+    if (!requestCacheService.isItemInCache(ProgrammeMembership.ENTITY_NAME, id)) {
+      log.info("Sending request for ProgrammeMembership [{}]", id);
 
       try {
-        requestCacheService.addItemToCache(ProgrammeMembership.ENTITY_NAME, uuidString,
-            dataRequestService.sendRequest(ProgrammeMembership.ENTITY_NAME,
-                Map.of("uuid", uuidString)));
+        requestCacheService.addItemToCache(ProgrammeMembership.ENTITY_NAME, id,
+            dataRequestService.sendRequest(ProgrammeMembership.ENTITY_NAME, Map.of("id", id)));
       } catch (JsonProcessingException e) {
         log.error("Error while trying to request a ProgrammeMembership", e);
       }
     } else {
-      log.debug("Already requested ProgrammeMembership [{}].", uuidString);
+      log.debug("Already requested ProgrammeMembership [{}].", id);
     }
   }
 }
